refactor launch profile command
