+++ src/commands/index.ts
@@ -5,6 +5,7 @@ import { deleteProfileCommand } from "./delete-profile"
 import { refreshProfilesCommand } from "./refresh-profiles"
 import { renameProfileCommand } from "./rename-profile"
 import { selectProfileCommand } from "./select-profile"
+import { launchProfileCommand } from './launch-profile'
 
 export const commands = [
     createProfileCommand,
@@ -13,6 +14,7 @@ export const commands = [
     cloneProfileCommand,
     deleteProfileCommand,
     refreshProfilesCommand,
-    selectProfileCommand
+    selectProfileCommand,
+    launchProfileCommand
 ]
